content/media/webaudio/test/test_bug938022.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test audio element currentTime is correct when used as media source</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var context = new AudioContext();
  var audio = new Audio("small-shot.ogg");
  audio.load();
  audio.addEventListener("loadeddata", function (e) {
    is(this.currentTime, 0, "currentTime must be 0");
    SimpleTest.finish();
  });
  audio.play();

  var source = context.createMediaElementSource(audio);
  source.connect(context.destination);
});
</script>
</pre>
</body>
</html>
